+++ wke/wke.cpp
@@ -39,6 +39,40 @@ void wkeInitialize()
     //WebCore::ResourceHandleManager::sharedInstance()->setCookieJarFileName("cookie.txt");
 }
 
+void wkeSetProxy(const wkeProxy& proxy)
+{
+    WebCore::ResourceHandleManager::ProxyType proxyType = WebCore::ResourceHandleManager::HTTP;
+    String hostname;
+    String username;
+    String password;
+
+    if (proxy.hostname[0] != 0 && proxy.type >= WKE_PROXY_HTTP && proxy.type <= WKE_PROXY_SOCKS5HOSTNAME)
+    {
+        switch (proxy.type)
+        {
+        case WKE_PROXY_HTTP:           proxyType = WebCore::ResourceHandleManager::HTTP; break;
+        case WKE_PROXY_SOCKS4:         proxyType = WebCore::ResourceHandleManager::Socks4; break;
+        case WKE_PROXY_SOCKS4A:        proxyType = WebCore::ResourceHandleManager::Socks4A; break;
+        case WKE_PROXY_SOCKS5:         proxyType = WebCore::ResourceHandleManager::Socks5; break;
+        case WKE_PROXY_SOCKS5HOSTNAME: proxyType = WebCore::ResourceHandleManager::Socks5Hostname; break;
+        }
+
+        hostname = String::fromUTF8(proxy.hostname);
+        username = String::fromUTF8(proxy.username);
+        password = String::fromUTF8(proxy.password);
+    }
+
+    WebCore::ResourceHandleManager::sharedInstance()->setProxyInfo(hostname, proxy.port, proxyType, username, password);
+}
+
+void wkeInitializeEx(const wkeSettings* settings)
+{
+    wkeInitialize();
+
+    if (settings->mask & WKE_SETTING_PROXY)
+        wkeSetProxy(settings->proxy);
+}
+
 void wkeFinalize()
 {
     wkeUpdate();